fix dtype in videomae pretrain

+++ src/smth2smth/shared/models/video_mae.py
@@ -434,12 +434,14 @@ class VideoMAEDecoder(nn.Module):
 
         tokens = self.proj(visible_tokens)  # (B, n_visible, d)
 
-        # Build full-length sequence: scatter visible tokens + mask tokens
-        full = self.mask_token.expand(B, N, -1).clone()
+        # Build full-length sequence: scatter visible tokens + mask tokens.
+        # Under AMP, ``tokens`` may be bfloat16 while ``mask_token`` is float32;
+        # ``scatter_`` requires matching dtypes.
+        full = self.mask_token.expand(B, N, -1).clone().to(dtype=tokens.dtype)
         idx_v = ids_keep.unsqueeze(-1).expand(-1, -1, tokens.shape[-1])
         full.scatter_(1, idx_v, tokens)
 
-        full = full + self.decoder_pos_embed
+        full = full + self.decoder_pos_embed.to(dtype=full.dtype)
 
         for block in self.blocks:
             full = block(full)
